moved lifi asset filtering into separate files
